backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
python-dotenv==1.0.0
aiofiles==23.2.1
pymongo==4.6.1
motor==3.3.2
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-generativeai>=0.8.0
anthropic>=0.39.0
kaggle>=1.6.0
datasets>=2.14.0
vaderSentiment==3.3.2
pandas>=2.0.0
autogluon.tabular>=1.1.0
scikit-learn>=1.3.0